Full working Trade Path finding algorithm!
FindTrades() currently returns a list of possible trade paths that can be taken to start with inventory items and arrive at the target item.

Total item count(s) in inventory are not taken into consideration / passed down through trades, inventory is considered static for this segment of the analysis, only serving to provide a baseline for trade paths to be considered.

Next step of implementation will involve creating a subsequent method to walk through each of the proposed trade paths and simulate them using the current inventory to establish viability / efficiency of each path