Initial PathListAnalysis algorithm, somewhat functional but with ONE MASSIVE FLAW / BUG
Currently this implementation of the Cost() method works decently well, recurssively working through each path and procurring the quantity of each common denominator that the overall trade path will cost. However, there is a massive oversight / problem with the current approach as it stands.

Currently each cost item for each Craft trade is evaluated individually, which will be fine in many cases, but will NOT be fine in others.

For example, the case of trying to trade F1 grenade to procure 1 Pistol Bullet....

In practice, this should result in a F1 cost of 1 SINGLE F1 grenade, as recycling the F1 grenade yields both the gunpowder  and metal fragments to craft a single set of pistol bullets.

HOWEVER, due to how the algorithm currently works, it will look at the gunpowder and metal fragments seperately, and conclude that a single trade is needed to procure both of them, FOR EACH INGREDIENT, effectively doubling the expected output, and presenting a cost of 2 F1 grenades instead of 1.

In order to address this, I will need to somehow implement some further set of checks to identify redundancies in steps, but this may be tricky, as redundancies may not always appear in the same step.

In reality, I may have to rework a lot of this from close to the ground up to account for these cases. Hopefully this wasn't an entire waste of time. monkaE